Removing signal handling from event loop
Because the worker event loop listens for signals, it is difficult for a
user of the library to call worker.Shutdown() on signal input as this
results in a panic from Shutdown being called twice. One can avoid the
panic by not calling Shutdown and letting the library handle it but then
it becomes hard to know when the program should exit and requires the
addition of a time.Sleep.

Letting the user of the library control what happens on signal input
give a greater level of flexibility.

Signed-off-by: Ivan Antolic-Soban <iantolic-soban@pindropsecurity.com>