Namespacing the checkpointer keys by stream name
The default limit in AWS on the number of DynamoDB tables is 256 and in
an effort to limit the number of tables created, it would be desirable
to allow the library to use the same DynamoDB table for N number of
workers. This requires namespacing the checkpointer table keys. We have
chosen to do this by creating the partition key for the checkpointer
table from both the stream name and the shard id.

Signed-off-by: Ivan Antolic-Soban <iantolic-soban@pindropsecurity.com>